moved apps list from page to modal, nav chore
