+++ src/apps/index.ts
@@ -3,24 +3,27 @@ import i18next from 'i18next';
 // types
 import { AppManifest, RecordPerKey } from '../types';
 
-const allowedApps = [
+export const allowedApps = [
   'sign-message',
 ];
 
-export const loadAppsList = () => {
+export const loadApp = (appId: string) => {
+  // eslint-disable-next-line @typescript-eslint/no-var-requires
+  const appManifest = require(`./${appId}/manifest.json`) as AppManifest;
+
+  if (appManifest.translations) {
+    Object.keys(appManifest.translations).forEach((languageKey) => {
+      const translation = appManifest.translations[languageKey];
+      i18next.addResourceBundle(languageKey, `app:${appId}`, translation);
+    });
+  }
+
+  return appManifest;
+}
+
+export const loadApps = () => {
  return allowedApps.reduce((apps: RecordPerKey<AppManifest>, appId: string) => {
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const appManifest = require(`./${appId}/manifest.json`) as AppManifest;
-
-    if (appManifest.translations) {
-      Object.keys(appManifest.translations).forEach((languageKey) => {
-        const translation = appManifest.translations[languageKey];
-        i18next.addResourceBundle(languageKey, `app:${appId}`, translation);
-      });
-    }
-
-    apps[appId] = appManifest;
-
+    apps[appId] = loadApp(appId);
     return apps;
   }, {});
 }
