+++ src/navigation/index.tsx
@@ -3,26 +3,26 @@ import { Route, Routes } from 'react-router-dom';
 
 // pages
 import Lobby from '../pages/Lobby';
-import Transfer from '../pages/Transfer';
-import History from '../pages/History';
 import NotFound from '../pages/NotFound';
-import Apps from '../pages/Apps';
 import Login from '../pages/Login';
+
+// apps
+import { allowedApps } from '../apps';
+
+// types
+import App from '../pages/App';
 
 export const navigationRoute = {
   home: '/',
-  transfer: '/transfer',
-  history: '/history',
-  apps: '/apps',
 }
 
 export const AuthorizedNavigation = () => {
   return (
     <Routes>
       <Route path={navigationRoute.home} element={<Lobby />} />
-      <Route path={navigationRoute.transfer} element={<Transfer />} />
-      <Route path={navigationRoute.history} element={<History />} />
-      <Route path={navigationRoute.apps + '/:appId?'} element={<Apps />} />
+      {allowedApps.map((appId) => (
+        <Route key={appId} path={'/' + appId} element={<App id={appId} />} />
+      ))}
       <Route path="*" element={<NotFound />} />
     </Routes>
   );
@@ -31,7 +31,7 @@ export const AuthorizedNavigation = () => {
 export const UnauthorizedNavigation = () => {
   return (
     <Routes>
-      <Route path={navigationRoute.home} element={<Login />} />
+      <Route path={'/'} element={<Login />} />
       <Route path="*" element={<NotFound />} />
     </Routes>
   )
