--- src/pages/History.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { useTranslation } from 'react-i18next';
-
-// components
-import { PrimaryTitle } from '../components/Text/Title';
-
-const History = () => {
-  const [t] = useTranslation();
-
-  return (
-    <PrimaryTitle>{t`title.history`}</PrimaryTitle>
-  );
-}
-
-export default History;
